% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/addDBs-methods.R
\docType{methods}
\name{addDBs}
\alias{addDBs}
\alias{addDBs}
\alias{addDBs,Experiment-method}
\title{Add new DBs at your experiment}
\usage{
addDBs(object)

\S4method{addDBs}{Experiment}(object)
}
\arguments{
\item{object}{The object for which we want modification.}
}
\value{
The object with the new databases which have been added.

The object with the new dbs
}
\description{
Add new DBs at your experiment
}
\examples{
exp <- new(Class="Experiment",
name="test",
date=Sys.Date(),
databases=list(1),
others=list())
listGenesIds <- list()
exp <- addDBs(exp)
}
