% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbInformations-functions.R
\name{changeDBNameIntoNumber}
\alias{changeDBNameIntoNumber}
\title{Function to know the number of the database associate to the name}
\usage{
changeDBNameIntoNumber(dbName)
}
\arguments{
\item{dbName}{the name of the db}
}
\value{
return the number of databases availables
}
\description{
this function return the number of databases availables
}
