% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/createNewDB-methods.R
\docType{methods}
\name{createNewDB}
\alias{createNewDB}
\alias{createNewDB,Experiment-method}
\alias{exp}
\alias{<-}
\alias{new(Class="Experiment",}
\alias{name="test",}
\alias{date=Sys.Date(),}
\alias{databases=list(1),}
\alias{others=list())}
\alias{createNewDB(exp)}
\title{Add manually a new database read by a file}
\usage{
createNewDB(object)

\S4method{createNewDB}{Experiment}(object)
}
\arguments{
\item{object}{The object for which we want modification.}
}
\value{
The object with the new database which has been added.

The object with the new db
}
\description{
Add manually a new database read by a file
}
