% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createNewDB-methods.R
\name{matchGenes}
\alias{matchGenes}
\title{matchGenes}
\usage{
matchGenes(gene, typeID, newListGenes, newAttributes)
}
\arguments{
\item{gene}{Gene}

\item{typeID}{id is msu(1) or rapdb(2)}

\item{newListGenes}{list of genes}

\item{newAttributes}{list of the attributes}
}
\value{
it will return a gene
}
\description{
Return the genes of the list which matches the gene
}
